Dockerfile
Oof, I do like deno, but I had discovered it part way through this project
and so the assets build is this weird frankenstein of npm/node + deno.
Oh well, this docker image should at least make it so I don't need to
worry about it too much when building in the future.